Clearing three tables' values to allow adding of the config table.
A foreignkey with a default targed id of 1 cannot be added as there's no ID=1
in a freshly created table, yet. Probably only a problem if there's already
data in the database, not for fresh ones.

I clear three tables to allow the migration to pass. Remco said the first two
(rainvalue and completerainvalue) are fine to delete, I guess the third is OK,
too.

The clearing happens in a transaction as you get a "pending trigger events"
postgres error otherwise.